+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Add `HybridNote` for recording species hybridization information
 - Add `SpecimenRepository` for storage `HybridNote` and **_specimen_**
 - Add `CaffeineComponent` for storage `caffeine` count
+- Add `SugarComponent` for storage `sugar` count
 
 ### Change